Stop prepending a zero byte to the input of hid_write()
hid_write() expects a report ID in the first byte.  The 0x11 and 0x12
values sent as the first byte of commands are consistent with report
IDs published in the keyboard's USB descriptors.

This change was motivated by commands not working reliably using
hidapi on macOS with a G Pro keyboard.