Rely on uaccess to control device access
The udev rules currently make supported device nodes world-readable
and writable, which means that any process on the system can read
traffic from keyboards including passwords etc. To avoid this, while
still allowing the "controlling" user to run g810-led without being
root, this patch adds a uaccess tag; this ensures that the user at the
console has write access to the devices. The mode is also changed to
660 to ensure that existing device nodes are fixed on upgrade.

Thanks to Xavi Drudis Ferran for bringing this to my attention.

Fixes: #293
Signed-off-by: Stephen Kitt <steve@sk2.org>